Feat. add useGetAuthStatus
